requirements.txt
streamlit==1.22.0
Pillow==10.0.0
tensorflow==2.14.0
requests==2.31.0
datasets==2.13.1
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.40.2
transformers==4.31.0
trl==0.4.7
flask==2.3.2